image-corruption/requirements.txt
opencv-python
scikit-image
brisque 
pillow